+++ custom_components/houseplan/frontend/houseplan-card.js
@@ -1,1 +1,1 @@
-globalThis.__HOUSEPLAN_BUILD_FINGERPRINT__="e499c3cfebb07203abd2397b1303ed1a45286f24fcc9129f45917b8a99ba1479";export{d3 as HouseplanCard}from"./houseplan-assets/houseplan-card-OQuczTJf.js";
+globalThis.__HOUSEPLAN_BUILD_FINGERPRINT__="fc2e76b2329cb102665073952b1b9f6346ff98d02332e2a07eb58f45429dc757";export{d6 as HouseplanCard}from"./houseplan-assets/houseplan-card-r12Gbmbf.js";